+++ data/src/main/java/com/agileburo/anytype/data/auth/mapper/MapperExtension.kt
@@ -396,7 +396,8 @@ fun Command.CreateDocument.toEntity() = CommandEntity.CreateDocument(
     context = context,
     target = target,
     prototype = prototype.toEntity(),
-    position = position.toEntity()
+    position = position.toEntity(),
+    emoji = emoji
 )
 
 fun Command.Replace.toEntity() = CommandEntity.Replace(