+++ data/src/main/java/com/agileburo/anytype/data/auth/repo/block/BlockDataStore.kt
@@ -26,7 +26,7 @@ interface BlockDataStore {
     suspend fun move(command: CommandEntity.Move): PayloadEntity
     suspend fun unlink(command: CommandEntity.Unlink): PayloadEntity
     suspend fun getConfig(): ConfigEntity
-    suspend fun createPage(parentId: String): String
+    suspend fun createPage(parentId: String, emoji: String?): String
     suspend fun openPage(id: String): PayloadEntity
     suspend fun openProfile(id: String): PayloadEntity
     suspend fun closePage(id: String)