+++ data/src/main/java/com/agileburo/anytype/data/auth/repo/block/BlockRemote.kt
@@ -24,7 +24,7 @@ interface BlockRemote {
     suspend fun updateCheckbox(command: CommandEntity.UpdateCheckbox)
     suspend fun move(command: CommandEntity.Move): PayloadEntity
     suspend fun getConfig(): ConfigEntity
-    suspend fun createPage(parentId: String): String
+    suspend fun createPage(parentId: String, emoji: String?): String
     suspend fun openPage(id: String): PayloadEntity
     suspend fun openProfile(id: String): PayloadEntity
     suspend fun closePage(id: String)