+++ data/src/main/java/com/agileburo/anytype/data/auth/repo/block/BlockRemoteDataStore.kt
@@ -16,7 +16,8 @@ class BlockRemoteDataStore(private val remote: BlockRemote) : BlockDataStore {
         remote.closeDashboard(id = id)
     }
 
-    override suspend fun createPage(parentId: String): String = remote.createPage(parentId)
+    override suspend fun createPage(parentId: String, emoji: String?): String =
+        remote.createPage(parentId, emoji)
     override suspend fun openPage(id: String): PayloadEntity = remote.openPage(id)
     override suspend fun openProfile(id: String): PayloadEntity = remote.openProfile(id)