+++ domain/src/main/java/com/agileburo/anytype/domain/icon/DocumentEmojiIconProvider.kt
@@ -1,0 +1,5 @@
+package com.agileburo.anytype.domain.icon
+
+interface DocumentEmojiIconProvider {
+    fun random(): String
+}