+++ library-emojifier/src/main/java/com/agileburo/anytype/emojifier/data/DefaultDocumentEmojiIconProvider.kt
@@ -1,0 +1,24 @@
+package com.agileburo.anytype.emojifier.data
+
+import com.agileburo.anytype.domain.icon.DocumentEmojiIconProvider
+
+class DefaultDocumentEmojiIconProvider : DocumentEmojiIconProvider {
+
+    companion object {
+        val DOCUMENT_SET = listOf(
+            "🌳", "⛳", "🧬", "🎈", "🎓",
+            "💡", "🎒", "🚀", "🤖", "📚",
+            "🍎", "🏡", "🤝", "😍", "☕",
+            "🔥", "💥", "✍", "⏳", "📌",
+            "🚩", "🦉", "📮", "📄", "🖌",
+            "🗳", "⏰", "🔑", "🎉", "🗃",
+            "🔖", "🧠", "👁", "🎗", "🎲",
+            "🧩", "🚲", "⚙", "🔶", "🌍",
+            "🏕", "🧳", "🌵", "🚗", "🚂",
+            "🖼", "⭐", "🥁", "🚠", "🛫",
+            "🏔", "🏗", "🛠", "🔍", "🕹"
+        )
+    }
+
+    override fun random(): String = DOCUMENT_SET.random()
+}